+++ src/Mirakl/MMP/Common/Domain/Message/Thread/ThreadRecipient.php
@@ -1,0 +1,13 @@
+<?php
+namespace Mirakl\MMP\Common\Domain\Message\Thread;
+
+use Mirakl\Core\Domain\MiraklObject;
+
+/**
+ * @method  string  getId()
+ * @method  $this   setId(string $id)
+ * @method  string  getType()
+ * @method  $this   setType(string $type)
+ */
+class ThreadRecipient extends MiraklObject
+{}