+++ src/Mirakl/MMP/Common/Domain/Message/Thread/ThreadReplyCreated.php
@@ -1,0 +1,13 @@
+<?php
+namespace Mirakl\MMP\Common\Domain\Message\Thread;
+
+use Mirakl\Core\Domain\MiraklObject;
+
+/**
+ * @method  string  getMessageId()
+ * @method  $this   setMessageId(string $messageId)
+ * @method  string  getThreadId()
+ * @method  $this   setThreadId(string $threadId)
+ */
+class ThreadReplyCreated extends MiraklObject
+{}