+++ src/Mirakl/MMP/Common/Domain/Message/Thread/ThreadTopic.php
@@ -1,0 +1,13 @@
+<?php
+namespace Mirakl\MMP\Common\Domain\Message\Thread;
+
+use Mirakl\Core\Domain\MiraklObject;
+
+/**
+ * @method  string  getType()
+ * @method  $this   setType(string $type)
+ * @method  string  getValue()
+ * @method  $this   setValue(string $value)
+ */
+class ThreadTopic extends MiraklObject
+{}